// src/types/Iphone13ConceptGLTFResult.ts
import { GLTF } from "three-stdlib";

export type Iphone13ConceptGLTFResult = GLTF & {
  nodes: {
    BackCover_Blue_0: THREE.Mesh;
    Screen_Screen_0: THREE.Mesh;
    CameraModuleBlack_BlackGlossy_0: THREE.Mesh;
    CameraModuleBlack_SpeakerAndMiic_0: THREE.Mesh;
    Bezel_BezelAndNotch_0: THREE.Mesh;
    Bezel_SpeakerAndMiic_0: THREE.Mesh;
    Bezel_CameraGray_0: THREE.Mesh;
    Bezel_FrontCameraGlass_0: THREE.Mesh;
    PowerButton_BlueGlossy_0: THREE.Mesh;
    Volume_Button_BlueGlossy_0: THREE.Mesh;
    MuteButton_BlueGlossy_0: THREE.Mesh;
    Sphere_Lens_0: THREE.Mesh;
    Camera1_CameraBlack_0: THREE.Mesh;
    Camera1_CameraMetal_0: THREE.Mesh;
    Camera1_CameraGray_0: THREE.Mesh;
    Sphere002_Lens_0: THREE.Mesh;
    Camera2_CameraBlack_0: THREE.Mesh;
    Camera2_CameraMetal_0: THREE.Mesh;
    Camera2_CameraGray_0: THREE.Mesh;
    Sphere001_Lens_0: THREE.Mesh;
    Camera3_CameraBlack_0: THREE.Mesh;
    Camera3_CameraMetal_0: THREE.Mesh;
    Camera3_CameraGray_0: THREE.Mesh;
    LiDar_LiDar_0: THREE.Mesh;
    Flash_Silver_0: THREE.Mesh;
    Flash_Flash_0: THREE.Mesh;
    CameraModuleGlass_Glass_0: THREE.Mesh;
    FrontCam_Lens_0: THREE.Mesh;
    Apple_Logo_AppleLogo_0: THREE.Mesh;
    IPHONE13_BlueGlossy_0: THREE.Mesh;
    IPHONE13_Red_0: THREE.Mesh;
    IPHONE13_BlueMatte_0: THREE.Mesh;
    IPHONE13_SpeakerAndMiic_0: THREE.Mesh;
    IPHONE13_Silver001_0: THREE.Mesh;
  };
  materials: {
    Blue: THREE.MeshStandardMaterial;
    Screen: THREE.MeshStandardMaterial;
    BlackGlossy: THREE.MeshStandardMaterial;
    SpeakerAndMiic: THREE.MeshStandardMaterial;
    BezelAndNotch: THREE.MeshStandardMaterial;
    CameraGray: THREE.MeshStandardMaterial;
    FrontCameraGlass: THREE.MeshStandardMaterial;
    BlueGlossy: THREE.MeshStandardMaterial;
    Lens: THREE.MeshStandardMaterial;
    CameraBlack: THREE.MeshStandardMaterial;
    CameraMetal: THREE.MeshStandardMaterial;
    LiDar: THREE.MeshStandardMaterial;
    Silver: THREE.MeshStandardMaterial;
    Flash: THREE.MeshStandardMaterial;
    Glass: THREE.MeshStandardMaterial;
    AppleLogo: THREE.MeshStandardMaterial;
    material: THREE.MeshStandardMaterial;
    BlueMatte: THREE.MeshStandardMaterial;
    ["Silver.001"]: THREE.MeshStandardMaterial;
  };
};